MovieChatBot/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
build==1.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
distro==1.9.0
Django==5.2.10
durationpy==0.10
faiss-cpu==1.13.2
fastapi==0.128.0
filelock==3.20.3
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.1.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langchain==1.2.4
langchain-chroma==0.2.3
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langchain-upstage==0.7.5
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
markdown-it-py==4.0.0
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.15.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
orjson==3.11.5
ormsgpack==1.12.1
overrides==7.7.0
packaging==25.0
pillow==12.1.0
posthog==5.4.0
propcache==0.4.1
protobuf==6.33.4
pybase64==1.4.3
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pypdf==4.3.1
PyPika==0.50.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
sqlparse==0.5.5
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.20.3
tqdm==4.67.1
typer==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.13.0
uvicorn==0.40.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
